Fix showing filters modal on mobile
Moved the search_facets block to where it's expected so that it renders the "show filters" button.
Updated the list to hide and modal to show at a larger break point so that there wasn't an in between state of having the filters column underneath results.
Tweaked style of modal to work with our new nav etc - not super pretty but usable at least.
Completely hiding the map filter if you're on a smaller screen. Alternatively we could just leave it at the bottom of the page (by removing the `display: none;`), because it scales down and looks/works alright, just isn't with the other filters in the modal.